updates fact tables to default to not truncating

+++ airflow/operators/load_fact.py
@@ -11,19 +11,22 @@ class LoadFactOperator(BaseOperator):
                  table='',
                  redshift_conn_id='',
                  query='',
+                 truncate=False,
                  *args, **kwargs):
 
         super(LoadFactOperator, self).__init__(*args, **kwargs)
-        self.table=table
+        self.table = table
         self.redshift_conn_id = redshift_conn_id
-        self.query=query
+        self.query = query
+        self.truncate = truncate
 
     def execute(self, context):
         self.log.info('Grabbing Postgres Hook')
         redshift = PostgresHook(postgres_conn_id=self.redshift_conn_id)
-        
-        self.log.info(f'Dropping data')
-        redshift.run(f"DELETE FROM {self.table}")
+
+        if self.truncate:
+            self.log.info(f'Dropping data')
+            redshift.run(f"DELETE FROM {self.table}")
         
         self.log.info(f'Running query {self.query}')
         redshift.run(self.query)
